+++ bottleship/__init__.py
@@ -3,6 +3,6 @@
 __author__ = 'Oscar Martinez'
 __email__ = 'omtinez@gmail.com'
 __version__ = '0.2.2'
-__all__ = ['BottleShip', 'data_encode', 'data_decode', 'data_is_encoded', 'parse_auth', 'static_file', 'redirect']
+__all__ = ['BottleShip', 'data_encode', 'data_decode', 'data_is_encoded', 'parse_auth', 'static_file', 'redirect', 'request', 'HTTPResponse']
 
-from .bottleship import BottleShip, data_encode, data_decode, data_is_encoded, parse_auth, static_file, redirect
+from .bottleship import BottleShip, data_encode, data_decode, data_is_encoded, parse_auth, static_file, redirect, request, HTTPResponse